apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "fastapi[all]>=0.121.1",
    "google-genai>=1.50.0",
    "groq>=0.34.1",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "websockets>=13.1",
]
